refactor(root-error): `Error` 페이지 컴포넌트를 이름 변경한다
